backend/requirements.txt
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.7
langchain-core==1.2.7
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.7
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.10.1
PyMySQL==1.1.2
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==6.4.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unidiff==0.7.5
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
